Create  Level 12 → Level 13
Level Goal
The password for the next level is stored in the file data.txt, which is a hexdump of a file that has been repeatedly compressed. For this level it may be useful to create a directory under /tmp in which you can work using mkdir. For example: mkdir /tmp/myname123. Then copy the datafile using cp, and rename it using mv (read the manpages!)

Username:ssh -p 2220 bandit13@bandit.labs.overthewire.org
Password:8ZjyCRiBWFYkneahHwxCv3wb2a1ORpYL